+++ docs/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.19] - 2026-07-26
+
+### Changed
+
+- Moved the queue-download retry and album-art defaults from the `miscellaneous` section
+  to the `downloads` section of the configuration file
+
+
 ## [0.0.18] - 2026-07-26
 
 ### Added